requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.23
python-multipart>=0.0.9
pdfplumber>=0.11
python-docx>=1.1
pandas>=2.0
spacy>=3.7
